pyproject.toml
[tool.poetry]
name = "iodev-tempclient"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
opencv-python = "^4.10.0.84"
PyYAML = "^6.0.2"
paho-mqtt = "^2.1.0"
pyqt5 = "^5.15.11"
watchdog = "^5.0.0"
ultralytics = "^8.2.82"
coloredlogs = "^15.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
